refactor(legacy): move version.ts out of legacy/ to src/
VERSION is a shared constant used by both cli.ts and http/status.ts —
it doesn't belong in the transitional legacy/ directory.
